Make balance unlocking more robust
Address audit issues 7.2 and 7.3.

Change public `updateBalancesUntil()` to internal
`_updateBalanceUntil()` and instead let updateBalances()
automatically catch up if necessary. In addition, add a test that
we can call updateBalances() even if many lifetimes have passed
since the last one.

Before, calls to updateBalances() within a day (24 hours) of
the last update would not change the balances, but would update
lastLockedTime, causing tokens to never get unlocked if this
happened repeatedly. This modifies the update logic to only
unlock tokens after full days so that the lastLockedTime only
updates when tokens are actually unlocked.