Enforce that requests in slates are current
Address audit issue 7.4.

Before, it was possible to include requests from previous epochs
in a slate. These requests might already be approved or expired,
and therefore not able to be executed. To avoid this, store the
epoch number associated with a request, and validate that it was
created in the current epoch when creating a slate.